feat: crud api 정리

+++ src/app/api/product-nutrition-100gs/route.ts
@@ -3,12 +3,12 @@ import { v4 as uuidv4 } from "uuid";
 import createHttpError from "http-errors";
 import prisma from "@/lib/prisma";
 import { handleError } from "@/utils/errorHandler";
-import { ProductNutrition100gCreateInputSchema, ProductNutrition100gSchema } from "@/types/schema";
+import { idSchema, ProductNutrition100gCreateSchema, ProductNutrition100gUpdateSchema } from "@/types/schema";
 
 export async function GET() {
   try {
-    const productNutritions = await prisma.productNutrition100g.findMany();
-    return NextResponse.json(productNutritions);
+    const nutrition100gs = await prisma.productNutrition100g.findMany();
+    return NextResponse.json(nutrition100gs);
   } catch (error) {
     return handleError(error);
   }
@@ -17,32 +17,32 @@ export async function GET() {
 export async function POST(request: Request) {
   try {
     const body = await request.json();
-    const parsedBody = ProductNutrition100gCreateInputSchema.parse(body);
+    const parsedBody = ProductNutrition100gCreateSchema.parse(body);
 
-    const newProductNutrition = await prisma.productNutrition100g.create({
+    const newNutrition100g = await prisma.productNutrition100g.create({
       data: {
         id: uuidv4(),
         ...parsedBody,
       },
     });
 
-    return NextResponse.json(newProductNutrition, { status: 201 });
+    return NextResponse.json(newNutrition100g, { status: 201 });
   } catch (error) {
     return handleError(error);
   }
 }
 
-export async function PUT(request: Request) {
+export async function PATCH(request: Request) {
   try {
     const body = await request.json();
-    const parsedBody = ProductNutrition100gSchema.parse(body);
+    const parsedBody = ProductNutrition100gUpdateSchema.parse(body);
 
-    const updatedProductNutrition = await prisma.productNutrition100g.update({
+    const updatedNutrition100g = await prisma.productNutrition100g.update({
       where: { id: parsedBody.id },
       data: parsedBody,
     });
 
-    return NextResponse.json(updatedProductNutrition);
+    return NextResponse.json(updatedNutrition100g);
   } catch (error) {
     return handleError(error);
   }
@@ -53,15 +53,15 @@ export async function DELETE(request: Request) {
     const url = new URL(request.url);
     const id = url.searchParams.get("id");
 
-    if (!id) {
-      throw createHttpError.BadRequest();
+    if (!id || !idSchema.safeParse(id).success) {
+      throw createHttpError.BadRequest("Invalid or missing ID");
     }
 
     await prisma.productNutrition100g.delete({
       where: { id },
     });
 
-    return NextResponse.json({ message: "Product nutrition 100g deleted successfully" });
+    return NextResponse.json({ message: "Product Nutrition 100g deleted successfully" });
   } catch (error) {
     return handleError(error);
   }
